pacemaker: Rename pacemaker.h to pcmk.h.
Pacemaker already includes a public pacemaker.h which we are going to
start using, and gcc doesn't like two header files with the same name in
different locations.

And rename pacemaker.c to pcmk.c for consistency.
